+++ engine/src/vytal/core/modules/input/input.h
@@ -14,7 +14,7 @@ VT_API Bool input_module_unregister_event(const InputEventCode code, const Deleg
 VT_API Bool input_module_invoke_event(const InputEventCode code, const VoidPtr data);
 VT_API Bool input_module_process_key_pressed(const InputKeyCode code, const Bool pressed);
 VT_API Bool input_module_process_mouse_pressed(const InputMouseCode code, const Bool pressed);
-VT_API Bool input_module_process_mouse_moved(const UInt16 x, const UInt16 y);
+VT_API Bool input_module_process_mouse_moved(const Int32 x, const Int32 y);
 VT_API Bool input_module_process_mouse_scrolled(const Int8 scroll_value);
 
 VT_API VoidPtr input_module_get_state(void);
